+++ matrixOperations.h
@@ -3,7 +3,7 @@
 #define FINAL_PROJECT_MATRIXOPERATIONS_H
 
 double **matrixAllocation(int numOfPoints, int dim);
-double **matrixSum(double **firstMatrix,double **secondMatrix,int sign, double numOfPoints, double dim);
+double **matrixSum(double **firstMatrix,double **secondMatrix,int sign, int numOfPoints, int dim);
 void sortByEigenValues(double **matrix, int numOfPoints, int dim);
 void freeMatrix(double **matrix);
 void printMatrix(double **matrix, int numOfPoints, int dim);
